tests/submission/Microsoft/semantic/footer.html

<!doctype html>
<html>
 <head>
  <meta content="text/html; charset=utf-8" http-equiv="content-type" />
  <title>verify FOOTER is a HTMLElement</title>
  <script src="http://test.w3.org/html/tests/resources/testharness.js"></script>
  <script src="http://test.w3.org/html/tests/resources/testharnessreport.js"></script>
 </head>
<body>
<div id="log"></div>
 <article>
  <footer id="footer">
   <p>P in an article and footer</p>
  </footer>
 </article>
<script type="text/javascript">
 test(function()
 {
    assert_array_equals( "[object HTMLElement]",
                         document.getElementById("footer").toString() )
 })
</script>
</body>
</html>
